legacy/requirements.txt
faiss-cpu==1.7.2
farm-haystack==1.14.0
gradio==3.22.1
openai==0.27.0
python-dotenv==1.0.0
pdfminer.six